requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
python-dotenv>=1.0.0
transformers>=4.36.0
torch>=2.2.0
huggingface-hub>=0.20.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9  # Para conexión con PostgreSQL/Supabase
